src/oblique_games/requirements.txt
glcontext==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
moderngl==5.12.0 ; python_version >= "3.10" and python_version < "4.0"
pygame-ce==2.5.3 ; python_version >= "3.10" and python_version < "4.0"
pygame-gui==0.6.13 ; python_version >= "3.10" and python_version < "4.0"
python-i18n==0.3.9 ; python_version >= "3.10" and python_version < "4.0"
